+++ src/Lakona.Game.Server.Hotfix.Abstractions/HotfixReloadStatus.cs
@@ -2,6 +2,7 @@ namespace Lakona.Game.Server.Hotfix.Abstractions;
 
 public enum HotfixReloadStatus
 {
-    Succeeded,
-    Failed
+    Succeeded = 0,
+    Failed = 1,
+    SucceededWithWarnings = 2
 }